Create FULL-SCHEMA-ish-2016SEPT25.sql
This is a work in progress - an attempt to provide a single script to deploy the pesp-was8.5 application. It contains on the CMS tablespace.

It began with a copy of the CMS.sql provided to OpenTechStrategies by Andy LaMora and has been updated to include the data-fixes-11-19-13.sql. The changes were insert near the table to resolve some colliding data row issues. If you are reading this and you care, likely you are trying to sort out the delta between the legacy files and this file, and the way I've done this is not making that easy. I'm optimistic we'll sort this down to a single, FULL_INSTALL so you do not have that issue.

This is far from complete. There are a number of other delta files that (probably) need to be incorporated.