WIP - kludging things to try to deploy to Wildfly
- Overwrite the websphere build.{properties,xml} with the versions from
  the old jboss repo.
- Delete WebSphereTransactionLookup. What does it do? Figure that out
  later; for now, it doesn't compile due to missing websphere
  dependencies.
- Delete the reference to DBIdentityProviderDAOBean in the ejb-jar.xml.
  Wildfly was complaining that it "Could not determine type of
  ejb-local-ref java:comp/env/ejb/gov.medicaid.services.impl.RegistrationServiceBean/identityProvider".
  Probably this will mean that nobody can log in, but getting that error
  in the UI will be a success case, and something to fix later.
- Switch persistence back to the version in jboss. What's the
  difference? Figure that out later.

This version should build with ant:
    $ cd psm-app
    $ ant deploy

It depends on an installed (aka unzipped) Wildfly; I have 10.1.0-Final.

Once built in ant, it will tell you where it put the resulting EAR file.
Trying to add the EAR application to Wildfly through the Wildfly
management console fails with a number of errors, which I believe point
to needed Wildfly configuration changes.