Mysterious Markdown Magic to fix formatting
By removing two spaces from the end of a line, make what is obviously
the same paragraph in the source be the same paragraph in the HTML on
GitHub.  I have no idea if that's a Markdown thing in general, or just
an idiosyncracy of GitHub-flavored Markeddown.  I suppose experiments
would answer that question, but I've already spent long enough writing
this log message, and now there is no time left for experimentation.