Delete WebSphereTransactionLookup
This class was added to support an old version of WebSphere. According
to the forum thread linked in the class[1], it may not work on more
recent versions of WebSphere, and in any case our primary supported
platform will be WildFly. Without an installed version of WebSphere, we
cannot even compile this class due to missing dependencies, so delete
it. If we need to support WebSphere and are faced with the problem this
class solves again, hopefully this will be a helpful breadcrumb.

[1] https://forum.hibernate.org/viewtopic.php?f=1&t=992310