Remove legacy JBoss data source automation
Old versions of JBoss used to have a file-based API which allowed
creation of data sources by dropping an XML file into a preset location.
This API was removed. In addition, as part of the migration to
WebSphere, the data sources were switched from `no-tx-datasource` and
`local-tx-datasource` [1] to be XA data sources, which allow
transactions across databases and messaging.

[1] https://developer.jboss.org/wiki/ConfigDataSources