use PostgreSQL data sources in psm-app
The PSM app uses Hibernate to interact with its database. The first step
in switching from Oracle to PostgreSQL is to tell Hibernate to expect a
PostgreSQL database instead of an Oracle database.

Document how to configure WildFly to provide data sources that connect
to a PostgreSQL database.

Issue #1 Support PostgreSQL or other open source database