remove WildFly-incompatible optimization
A local reference is an optimization that tells the application server
that a bean is in the same container, which allows it to avoid a remote
call.

WildFly 10 seems to be having trouble with this particular local
reference, giving the error `Could not determine type of ejb-local-ref`.
After trying a number of variations of this element without success,
remove it. If it turns out to be a performance problem, we can spend
more time on it.