include built core file in lib directory
The `build-core` ant task builds `cms-core.jar`. The `dist` task used to
copy this file from the build directory to a lib directory, until
ddcabd6d0b942cf1aeed18fd2761a88d3d4e45c4. In that commit, I <include>d
it in the ear file, but it was put in the root of the zip, instead of in
`/lib`, where it is needed. Ant, or at least the version of ant I have
on my machine (1.9.6), doesn't seem make it easy to rename a file as
part of including it.

So, mostly revert back to the earlier behavior. Move the copy to the end
of the `build-core` task (previously it was in the `dist` task), and add
the copied jar to the list of things to delete in the `clean` task.