Install script for RHEL 7.3
I put together a shell script that will install all dependencies
and deploy the application without requiring user interactions.
The hope is that this can eventually be used to guide a Docker
version of the application.  In the short term this should also
help us more effectively test, improve, and deploy the
application.