Remove build artifacts
External sources had build artifacts checked in to the repo.  This
project has an unwritten guideline against putting files in the repo
that can be generated from other materials int he repo.  I ran
ant-clean and then committed the results.

The unwritten guideline should be written and added to
CONTRIBUTING.md.  I have written it and will submit a patch on it.