Add provider type seed data
Data from a previous system's database export.

Note that there are provider types defined in
`cms-business-model/src/main/resources/Entities.xsd` that are not
included in this seed data. Possibly all of this will need to be
replaced once we have a better understanding of how provider types work
and what our requirements are.