Switch enrollments over to annotations
Pluralize the table name, use a more descriptive column names, and
generate primary key values using Hibernate natively.

Note that the old table was named `TICKET`, and some of the fields and
comments reflect that legacy name, as well. This is confusing and should
be cleaned up later, but is out of scope at the moment - see issue #57.

Issue #36 Use Hibernate 5, instead of 4