Fix potential null pointer exception bug
The intent of the previous code seems to be to return all provider types
if applicant type is not set, but switching on a `null` causes a
`NullPointerException`. Adding null checks leads to confusing code, and
allowing null parameters leads to confusing APIs, so instead create a
separate method that returns all provider types and call that.

It's not clear to me when this code is invoked; possibly it's on or
behind a page that has other errors that prevent me from loading it.