Debug force dupes
* Force dupes should be reload for UPDATED.csv - While improving the
test, I realized I had cast the force_dupes stuff wrong in the
exclusion etl.  We completely replace the table on each etl, so there
are no dupes possible.  The question is really whether we take the
time to reload the data or not.  I recast the option as force_reload.

* Reinstatements shouldn't be able to force dupes - Remove force_dupes
from reinstatement etl.  There isn't a scenario where we want dupes,
so just find them and prevent them.