Extract ProviderType-AgreementDocument link
Each provider type (eg: audiologist, x-ray services) may have a number
of documents associated that new providers of that type must agree to.

Previously, that relation was encoded on the application side as one of
the polymorphic entries in `ProviderTypeSetting`. The upside to this
approach is that it avoids the eager/lazy loading problem by giving
explicit control to the application as to when and how it is loaded, and
can avoid extra lookups when updating. When this was originally written,
I don't think there were any really great solutions to this problem.

There are some downsides to that approach, however. The most major one
is that it is not working with Hibernate 5 and PostgreSQL! Because of
the polymorphism, where one table has application-side foreign keys to
multiple other tables by a combination of (entity type, entity ID)
columns, IDs are stored as strings - because sometimes they link to
entities whose IDs are strings, and sometimes they link to entities
whose IDs are numbers. PostgreSQL doesn't like that: it complains that
there is no equality comparison operator that takes a string and a
number. More broadly, it takes foreign key integrity checking out of the
database, as well as making the schema more difficult to discover or
understand.

This, the ProviderType-AgreementDocument link, is the first that
manifested this problem. The solution is to make the relationship known
to Hibernate, and extract a new table:
`provider_type_agreement_documents`. This is encoded as a many-to-many
relationship from `ProviderType`. Telling Hibernate about the relation
allows us to lean on it to do the heavy lifting of finding / updating.

Many-to-many relationships are lazily loaded by default, meaning that
when the ORM loads an instance of `ProviderType`, it does not also load
the related `AgreementDocuments` until someone invokes the
`getAgreementDocuments` method.  This must be done in a session, or else
it will throw a `LazyInitializationException`. Views (JSPs) are rendered
outside of a session, so `ProviderType` instances must be loaded with
their related `AgreementDocuments`. We don't always want to do that,
though, because there is a performance cost to eager loading,
particularly for large objects (how big is the text of an
AgreementDocument? almost certainly more than we want to fetch and not
use).

Calling `getAgreementDocuments` is one way to initialize, and is
convenient, but is not always performant: you have to make a query to
get the `ProviderType`, and then you have to make another query to get
the `AgreementDocument`s. If you know in advance that you want both, you
can get both in the same query. JPA 2.1 introduced Entity Graphs[1] to
help with this problem.

According to the database dumps, there are more links hidden in
`ProviderTypeSetting`, and I expect future commits will address those in
a similar way.

Issue #36 Use Hibernate 5, instead of 4

[1] http://www.thoughts-on-java.org/jpa-21-entity-graph-part-1-named-entity/