Extract ProviderType-LicenseType link
Following the pattern of the ProviderType-AgreementDocument link, switch
from the application-side polymorphic link table backing
`ProviderTypeSetting` to a more specific link table that Hibernate
understands directly. This means creating a new EntityGraph to describe
a different partially-initialized ProviderType, and using it in a more
specific ProviderTypes lookup query.

The seed data is taken from `mpse-clean-setup.sql`. That data is also
somewhat duplicated in the drools rules; you can't create an acupuncture
enrollment without an acupuncture license, and that is enforced by
drools - even if the ProviderType is not associated with the LicenseType
in the database. (See issue #79.)

Issue #36 Use Hibernate 5, instead of 4