Add pull request review to contributing
We took an informal poll in channel.  James, Cecilia, and Slifty all
said we like the PR process as the strong default.  Karl bowed to the
majority. I picture him on his knees with a sword to his neck,
swearing vengeance under his breath.

This patch just writes down the result of that poll and makes it
policy.