List current PSM capabilities in README
I tested the PSM yesterday. Because of the ongoing Hibernate
migration, most PSM functionality is currently unavailable.
As we recover old functionality, I expect we will massively
expand this section of the README and discuss capabilities at
the level of abstraction that the old README did, rather than
focusing on the fairly trivial features that work right now.