Annotate Entity, Person, and Organization
`Person` and `Organization` extend `Entity`. Previously, this was
expressed in the database as a single table (`ENTITY`) with a
discriminator field; Hibernate would read that field and construct the
appropriate object (`Person` or `Organization`). This can have
performance benefits over the alternatives, but I find it makes the
schema a little harder to understand.

In migrating these for Hibernate 5, separate them into three distinct
tables using the joined inheritance strategy. The `people` and
`organizations` tables share a primary key with the `entities` table,
and add their subtype-specific fields. Hibernate figures out which kind
of `Entity` an entity is by joining against those two tables (that is,
we are omitting the optional discriminator column - if that turns out to
be a performance penalty we care about, we can add it after profiling).

See also: https://en.wikibooks.org/wiki/Java_Persistence/Inheritance

Issue #36 Use Hibernate 5, instead of 4