Fix lazy initialization bug with agreement docs
When I extracted the ProviderType-AgreementDocument link in
be9c9fe4a614f8162476260ca929306a3b496fe5, there were sections of code
that I could not execute because they were blocked by the incomplete
Hibernate migration. Now that more functionality is working, some of
those problems can start to be fixed.

In this case, we need to look up a ProviderType from a
ProviderInformationType, and we need the AgreementDocuments to be
loaded. Introduce a new method to look up a ProviderType by its
description and prefetch the associated AgreementDocuments.

Issue #36 Use Hibernate 5, instead of 4