Allow null enrollment status on Entity
The Hibernate XML mapping file said the `enrolled` field must be backed
by a non-nullable column, but looking at the source, there is no default
and it is only ever written to with the string 'Y'. Looking at the
Oracle database instance, it appears that the table was created with a
nullable column. Do the same.

We should fix this, long-term; this appears to be another artifact of
Oracle not having a boolean data type. See issue #57.

Bug introduced in 322664d4192e238d6960a62ab7022126a350086f.

Issue #36 Use Hibernate 5, instead of 4