Use generated ID and persist enrollment details
Before the migration, persisting an enrollment went something like this:

1. generate an ID for the enrollment
2. set the enrollment ID on the enrollment
3. persist the enrollment
4. persist the related (but not linked in Hibernate) entities, using the
   enrollment ID as an application-side foreign key

In converting to Hibernate 5, and using Hibernate-generated IDs, steps 2
and 3 are merged. In this case, step 4 needs to change somewhat.

`EntityManager#merge`[1] takes an unmanaged instance and returns a
managed instance.  When we were manually setting the ID, we didn't need
to use the managed instance, and could ignore the return value. Now that
we're using the Hibernate-generated ID, we need to get the new ID from
the return value, but the `@Transient ProviderProfile details` field
doesn't get copied from the unmanaged instance to the managed instance -
which, of course, is part of the point of `@Transient`.

Save the `details` before merging and use the ID from the merged entity.

[1] https://docs.oracle.com/javaee/7/api/javax/persistence/EntityManager.html#merge-T-

Issue #36 Use Hibernate 5, instead of 4