Enabled gradle build
This commit begins the process of removing ant in place of gradle.
Specifically, it adds support for gradle builds, though it does not
remove ant files.

Specifically it contains:

- The gradle wrapper
- The gradle build script and project settings
- Updated properties and application.xml files to use new build names
- Removed a few unused components of the original build

Ant builds will no longer deploy, as a result of those changed names.
The build script also needs to be updated to use gradle.

This commit does not yet begin the process of switching dependency
management.