Remove obsolete Oracle warning and mention RHEL script
The PSM now uses Postgres, so I believe the paragraph about
our future PostgreSQL plans is now obsolete. Removing.

As with the Docker install, we should refer to the RHEL script
in this document, and we should refer here to any other installation
approaches (such as via package managers) in the future.