Remove unused `verified` field from License
The field is not persisted in the database, and no code reads from or
writes to it. I confirmed that two ways: first, IntelliJ shows that
there were no Java usages; verifying no usages at all meant `git grep -i
verified` and looking through each use of the word in the JSPs and the
drools rules.

I haven't dug too deep, but I believe the workflow for approving an
enrollment (and verifying the attached license(s)) has everything in
memory - I don't think there is an in-between state persisted where the
license is verified but the enrollment is not.