Change INSTALL.md recommendation from Mailcatcher to Python SMTPD
Our early choice of Mailcatcher as a suggested email server for use by
PSM developers has caused people some installation problems, since
various Ruby-related dependencies are often not on users'
systems. Therefore I did some research and, after failing to find a
suitable Java-based replacement application, I suggest we go with this
Python one-liner. If we find an easy drop-in replacement for
Mailcatcher in the future, one that is easy to deploy with WildFly and
that gives the user a web view of sent messages, we should switch to
that instead.

Fixes: #113.