Set external sources URL to example domain
We've removed the source code to the external sources app, but there are
still references in the software. This one was pointing at an AWS IP
address; for now, redirect it to example.com. We'll need to remove this
when we remove the code that uses it.

Issue #11 Separate ext-srcs from main application