Mavenifiy portal services dependencies
This commit contains the first step of converting project dependencies to Maven.
Jars packaged as part of the cms-portal-services have been replaced with Maven
dependencies where possible.  There are a few custom TopCoder packages which still
remain and may need to be replaced in a more involved fashion.

This commit does not remove other packaged jars in other sub projects.