Manage jBPM dependencies in gradle
Remove the vendored jBPM and its dependencies' jars from the repo, and
manage them in gradle. gradle knows how to pull in transitive
dependencies, so we do not need to name them explicitly: we only need
the top-level dependencies that we refer to directly, in code or in
configuration.

Issue #16 Manage sets of dependencies via Gradle or another tool