Upgrade jBPM from 5.3 to 5.4
jBPM 5.3 doesn't play well with PostgreSQL and JPA 2:

> org.postgresql.util.PSQLException: Large Objects may not be used in
> auto-commit mode

This was reported in issue JBPM-3763[1] and fixed in jBPM 5.4. Upgrade
jBPM to 5.4, and its dependencies to the versions shipped with jBPM 5.4,
to address this issue.

Some APIs changed in the upgrade, and notably the persistence definition
of `ProcessInstanceInfo` was moved into a separate file which we must
now include in our `persistence.xml`.

The schema for jBPM's entities changed, so as part of the upgrade,
finally extract the schema into its own SQL file, make sure it has
sensible types and keys in PostgreSQL, and delete the old schema from
`legacy_seed.sql`. Update the docs to match.

[1] https://issues.jboss.org/browse/JBPM-3763

Issue #1 Support PostgreSQL or other open source database
Issue #171 Service administrator cannot view or approve pending
           enrollments