Completed removal of packaged FOSS jars
This commit completes the removal of packaged jars for any library
that exists in the maven repository.  It does not include custom jars
that are part of the project.  It also does not include jars packaged
in the data_load project, since that projet is potentially obsolete and
not currently part of the gradle build.