Replace placeholder footer text: #108
Remove "Companyname Lorem Ipsum" in the footer and replace with a link
to the repository.  Use the included footer in two files where it had
been set manually, so that we only need to change it in one place in the
future.

Remove the unused "date" variable, as @jasonaowen suggested.