List run-time dependencies with licenses: #147
Add list of run-time dependencies, with their licenses and
subdependencies.  These are the jars removed in 02292be and ed5f4d8,
which @slifty will convert to pull from the Maven repo, as discussed in
PR #193.  Several of those jars were duplicates of jars already covered
in this file, including almost all the jars in the `WebContent/WEB-INF`
subdirectory.  Those duplicates are not included in this list.