Added back potential runtime dependencies
During the jar purge the only libraries that were brought over to maven / gradle
were ones that were required for compile or initial deploy.  This was an effort
to remove unused dependencies while doing the transfer.  However, some of the
libraries were dynamically used during runtime, causing functionality to break.

Since the point of this PR is not to cull unused libs, but rather to simply change
how they are loaded, all libs have been moved to maven / gradle.  Removing unused
libs should be handled at a later date.