Remove FileNet service
From Wikipedia: "FileNet... is a framework for developing custom
enterprise systems, offering much functionality out of the box and
capable of being customized to manage a specific business process."

We do not have a FileNet instance to work with, nor do we know if the
states that are considering using this application use FileNet or want
to integrate with it. What we will want is a more modular system that
individual states can configure, one module of which might support
FileNet.

Delete the supporting code for FileNet, and leave comments behind to
remind us of where we should build integration hooks.

Issue #215 Add extension points to allow per-state integration