Remove outgoing messaging integration
It appears that there was once a service that listened for JMS messages
in a very specific format. We do not have access to that system, or know
anything about it; more importantly, we don't know if other states use
such a system.

These messages were sent when an enrollment was approved, but the queue
is not configured and sending the message fails, preventing enrollment
approval. Delete the message-sending code, and leave behind a comment
reminding us of where we should build integration hooks.

Issue #215 Add extension points to allow per-state integration