Remove incoming message integration
As PSM once emitted messages for a remote system to use to stay in sync,
so too did it once listen to messages from that remote system and update
profiles accordingly. Again, though, we do not have access to such a
system, and it is likely that other states have different integration
requirements.

Delete the message-driven bean and its now-unused dependencies.

There isn't really a place left behind in the code to add a comment
reminding us of integration points; it's also not 100% obvious to me
that we want a similar such incoming integration point - we may be
better served by an API, rather than a message bus.

Issue #215 Add extension points to allow per-state integration