Switch ProviderCategoryOfService to annotations
Rename the tables, rename the columns, and use Hibernate to generate
IDs.

Hibernate generates a query with left outer joins to look up categories
of service by provider.  We need a `SELECT DISTINCT` in the HQL query
because, without it, Hibernate returns multiple, duplicate entities when
there are multiple results per entity (that is, multiple categories of
service per provider). While this does slightly affect the generated
query, this is more of a hint to Hibernate on how to behave.

See also:
- https://developer.jboss.org/wiki/HibernateFAQ-AdvancedProblems#jive_content_id_Hibernate_does_not_return_distinct_results_for_a_query_with_outer_join_fetching_enabled_for_a_collection_even_if_I_use_the_distinct_keyword
- https://stackoverflow.com/q/27170142
- https://stackoverflow.com/q/1995080

P.S. I agree with this comment:
https://stackoverflow.com/q/1995080#comment5962237_1995244

Issue #36 Use Hibernate 5, instead of 4