Switch ProviderTypeSetting over to annotations
ProviderTypeSetting is more complicated than it looks; it's not just a
bag of text fields. Several different kinds of information are related
to ProviderTypes, such as:

- a dentist might specialize in orthodontics
- a physician might be licensed by either the ABMS or the ABPS

Before our migration to Hibernate 5, this class/entity/table was also
used to relate provider types to agreement documents (ie, an audiologist
must agree to Form A) and to license types (ie, an acupuncturist must
have an acupuncturist license); those were removed in
be9c9fe4a614f8162476260ca929306a3b496fe5 and
81e0bca0482726034dac297a6789c355f40af950. They needed to be removed
because the column containing the related entity's primary key value is
textual, those are numeric, and PostgreSQL won't allow you to join on
mismatched types.

However, all the remaining uses of this table relate to LookupEntities,
which have textual primary keys. I find this pattern not to be great
because it makes it much harder for me to understand how these various
types relate, since the relation is not really expressed in the
database.  However, it's working, and we can clean it up when it's a
higher priority. For now, mark the class and the helper method that uses
the class as deprecated.

Rename the table, rename the columns, and pull seed data from
mpse-clean-setup.sql. Adding this seed data should also address issues
relating to enrollment of a few specific provider types.

Issue #36 Use Hibernate 5, instead of 4
Issue #146 Dentist or physician requires a specialty, but no specialties
           appear
Issue #177 Personal Care Assistant Enrollment Application has no issuing
           states