Automatically build and publish Javadocs
I've adapted the approach explained in
https://benlimmer.com/2013/12/26/automatically-publish-javadoc-to-gh-pages-with-travis-ci/
to have Travis automatically build our Javadocs and push them to
the `gh-pages` branch, thus updating the GitHub Pages site
https://opentechstrategies.github.io/psm/javadoc/ , per #30 . This
depends on a GitHub repo owner creating a GitHub access token with
public_repo rights and then adding it to the Travis CI repo settings
twice, once with the name "GH_TOKEN" and once with the name "secure".
Cecilia has done so.

As part of this work I've also created a utility scripts directory
and moved the Red Hat install script there, per a suggestion from
Jason.