Converted dependencies to Gradle
This commit removes packaged jars that exist in maven repositories
and shifts their management over to Gradle.  There are a few jars
that appear to be TopCoder libraries, and do not appear in public
repositories.  Those jars are still bundled with the rest of the
code.

Some jars do not have direct mappings to gradle, as they are
transitively bundled through other dependencies.  Furthermore there
are a few libraries which were not required during the compile step
but are required at runtime.  The  list should be
reviewed in a separate PR to determine whether or not they are legacy
dependencies that are no longer used.