Throw exception on import profile
From reading the code, it appears to me that importing profiles from
external services does not work correctly. We can't test this, because
there are no unit tests or integration tests, and we don't have access
to such an external service or its specs.

It uses the legacy sequence generation that we're moving away from,
which is what drew my attention to it. Since the functionality seems not
to work, and we have no way to test it, and this is likely to change
once we start working on integration with states' systems, and it's in
the way... just delete it and throw a not implemented exception.

Issue #36 Use Hibernate 5, instead of 4
Issue #39 Make state-specific auth configurable
Issue #181 Provider cannot link profile with external auth