Reuse id for profileId
`ProviderProfile` uses `control_no` (the column backing the field `id`)
as its primary key, but it also has `profile_id`, which is only assigned
when an enrollment is approved.  Previously, the app was generating this
(non-primary-key) profile ID the same way as it was generating all the
other primary keys, but with the migration to Hibernate 5 that approach
no longer works.

@slifty suggested reusing the `control_no` value generated by Hibernate
as the `profile_id`, which is a good idea. This allows us to defer a
complete rework of ProviderProfile, avoids introducing any new entities
simply for the purpose of generating an ID, and is relatively
straightforward.

Issue #36 Use Hibernate 5, instead of 4