Create jBPM sequence for task comments
When adding the database schema for the new version of jBPM in
fdc29e552ef92c474b7dbf4d986b7b76607b307c, I missed one of the sequences.
This showed up as a warning on running the SQL schema creation multiple
times:

    NOTICE:  sequence "comment_id_seq" does not exist, skipping

Presumably there is also some jBPM behavior that is broken without this
which we have not yet exercised.

Issue #1 Support PostgreSQL or other open source database