Remove ServiceAssuranceExtType
This lookup entity had no lookup data (which was one of the issues with
the partially-implemented Chemical Dependency / Assured Services
feature), and is no longer referrenced. Delete the entity, as well as
the unused XML-defined type of the same name.

Issue #36 Use Hibernate 5, instead of 4
Pull request #93 Redo ServiceAssuranceExtType as annotations