Remove assured statement constants
The two page fragments, while never referenced from any UI contexts,
shared an enum value which was used to add the form when the provider
type was "Chemical Dependency Program". Since the page fragments were
unreferenced, and the form binder was unconfigured, it's a good thing
that the seed data for the ProviderType entity was also missing this
type!

Delete the form ID enum value, the provider type XML enum value, and the
logic using them.