Extract query helper method that fixes warning
There are two ways we look up a provider's categories of service: by
enrollment ID and by provider ID. The former is for enrollments that
have not yet been approved; the latter is for enrollments that have
been. These queries are very similar, and were originally written
without the benefit of typed queries, which necessitated suppressing a
generics warning.

Extract a helper method for the common parts of the query, and use a
typed query instead of an untyped query to eliminate the warning.