Remove lorem text on account setup: #108
The lorem ipsum text that showed up when a new provider logged in was
distracting.  Remove it.  I commented out the "Linked Profiles" section
on that same page, too, because that functionality isn't enabled right
now.

Also, fix #152 by adding a new comment-beginning mark to properly escape
the end of the comment and the high-risk helptext.