Travis integration tests enabled
Travis now properly start up the wildfly server and runs integration
tests using SauceLabs.  The documentation for properly setting up the
environment variables for saucelabs will come in a later commit.

The big trick to getting wildfly to start up seemed to be separating
the actual server launch from the app deployment, and putting the
server startup script in the `before_script` section.

This commit also changes the way uploads are handled in some of the
browser tests.  The previous upload method did not work when the
browser was remote.  By using serenity's upload method the tests
will properly run on a remote server.