Associate labels with inputs in tribal information
By nesting the <input>s inside the <label>s, browsers know that the two
are associated, and allow clicks on the text to count as clicks on the
radio button.

I verified that the computed CSS rules using Firebug had no meaningful
change - the only difference is that the radio buttons now have the
style `font-weight: normal`.