Fix scope warnings in BusinessProcessServiceBean
These methods are never used outside of this class, probably should not
be, and since users of this class typically have a reference to it
through the interface it implements, probably cannot use these methods.
IntelliJ knows that they are otherwise unused, and warns about
unnecessarily broad scope. Mark them as private.