Remove incorrect comment
The danger of comments that describe "what" rather than "why" is that,
not only are they redundant when they're first written, they are likely
to be missed when the "what" is changed. In this case, it appears that
the original upload size was five megabytes, and when it was discovered
that that was too low, the value was raised but the comment was not.

Issue #263 Address "Entity too large" error