Enable Continuous Deployment
Travis supports generic deployment scripts, although they note that the
feature is essentially in beta mode at this point.  There are some
challenges with using travis for deployment given the fact that this
is an open source project.  Specifically, using Travis as our
platform for deployment requires us to bake our deployment
infrastructure into the repository itself.

You can see the nature of the baking by looking at .travis.yml.
In particular the fact that there is a single deploy location is part
of the repository, and detailes about that server's configuration are
predefined by travis.

We may want to consider having Travis compile into Docker and having
some type of CD through Docker in future to avoid this.

NOTE: Right now we have encrypted server information as well as an
encrypted private key into the repository.  This will be removed and
replaced with env variables written in the interface soon.