Add script to extract CSV files from LEIE site
Written by Kevin Wan and originally submitted in
https://github.com/OpenTechStrategies/psm/pull/240 ,
and polished by Sumana Harihareswara.

This script doesn't have tests, and in ordinary operation,
users will rely on the integrated ETL process that starts
with etl/leie/leie/etl.py . This standalone script helps
if a user wants to download and inspect the LEIE files
as a spot check.