Start on multi-stage HTML/JSP comment fix (#108)
Remove a comment, because its opening "<!--" and closing "-->" didn't
match the div hierarchy, and because the way the comment tags were
inserted caused mis-indentation.  A subsequent commit will reinstate
the comment with proper balancing; I'm doing it as a separate commit
so that its relationship to the hierarchy will be easier to see from
the now-restored indentation.