Finish multi-stage HTML/JSP comment fix (#108)
Well, "multi-stage" turned out to be just "two-stage".  This commit
follows up to commit 610a6479 with a restructured comment that has the
intended effect of the comment originally added in commit f5d70d7bb9,
but preserving indentation and balancing this time.  This commit also
takes the portion of the commit f5d70d7bb9 log message that probably
should have been a comment in the code all along and makes it so.