Add and transclude contextual help file about NPIs
This is a first step towards #265. If the user hovers
their mouse pointer over the abbreviation "NPI" during
the Personal Info step of the enrollment process, they'll
see a short informational message. We need to polish the
presentation but this first step is useful in itself as
well as serving as a proof of concept.