Fix comma placement in confirmation screen
During provider enrollment, on the Step 4 ("Summary") screen, in the
"practice information" confirmation display, the final line of the
practice address should have a comma and then a space between the city
name and the state, and no comma between the state and the ZIP code.

The old behavior would omit a comma if the state field were empty.
Jason and I decided that, in the unlikely event that a user somehow
got to the end of the process without specifying a state, then having
an address with a trailing comma after the city was a reasonable
trade-off.

Since the city, state, and ZIP variables are only used once each in
these templates, Jason suggested we simplify the code by removing the
`c:set` assignments.

Fixes #159.