Remove application-side database ID generator
As part of the migration from Hibernate 4 to Hibernate 5, we changed how
all the entities received IDs. Previously, we generated those entity IDs
in the application, but the APIs used to do so were deprecated; now, we
annotate the IDs to tell Hibernate in its role as a JPA persistence
provider to generate IDs automatically.

The code that generated these IDs in the application has a bunch of
deprecation warnings, and is no longer used. Delete it.

Issue #36 Use Hibernate 5, instead of 4