Allow admin to view license attachments in review
Previously, trying to view an uploaded license would cause an error:

    org.springframework.web.HttpSessionRequiredException: Session
    attribute 'enrollment' required - not found in session

While @cecilia-donnelly and I both remember this working at one point,
if we are remembering correctly it was from a long time ago: I went back
about six weeks trying to figure out what change caused this, and I
wasn't able to find a working version.

The exception relates to the way the attachment was being retrieved. For
some reason, we were looking for the enrollment model in the session,
using that to get its MIME type and filename, and then looking it up in
the database. The database has the MIME type and filename, and the
existing method checks that the user has permission to view the
attachment, so stop looking for information in the session and use the
values from the database.