Upgrade Apache Commons FileUpload library to 1.3.3
Fix several security vulnerabilities present in our old version by
upgrading to the latest version.

We seem to use FileUpload purely by including it in our classpath - no
code refers to it, and no XML configuration mentions it, but omitting it
prevents our application from deploying due to a missing dependency.

I tested this by deploying it, creating an enrollment with a license
attachment as a provider, and viewing that attachment as an admin.

See also: Apache Commons FileUpload Security Vulnerabilities
https://commons.apache.org/proper/commons-fileupload/security-reports.html