Remove unneeded direct dependency on spring-asm
We do not directly refer to anything in the spring-asm jar, so we do not
need to depend on it directly. It is still part of the project because
other spring dependencies transitively depend on it, including
spring-context and spring-webmvc.

Aside from a conceptually simpler build, this should also ease upgrading
Spring, as it means one fewer thing to explicitly upgrade and test.

Issue #16 Manage sets of dependencies via Gradle
Issue #157 Review libraries for needed updates
Issue #219 Upgrade Spring to 3.2.18