Query LEIE API on new provider enrollment
Reimplement the excluded provider handler to query the LEIE API for the
provider's NPI on enrollment submission. The results, if any, are added
to the automatic screening log visible to the admin at enrollment review
time. If the provider's NPI is not found in the LEIE, the checkbox next
to "Check Excluded Provider List in OIG" will be checked.

Issue #289 Use LEIE API to automatically screen providers